fix : 자료형문제 test중
Former-commit-id: 293d0bb6e3cefc52fc97838ccb5cb7824c890f14

+++ backend/src/main/java/com/tfeo/backend/domain/home/model/entity/Home.java
@@ -46,7 +46,7 @@ public class Home {
 	private String hostName;
 
 	@Null
-	private float hostAge;
+	private Integer hostAge;
 
 	private String hostPhone;
